+++ lib/commands/commands.dart
@@ -1,3 +1,4 @@
 export 'create_command.dart';
 export 'help_command.dart';
 export 'connect_command.dart';
+export 'add_command.dart';