+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+## 0.2.0
+- added possibility to create folder structure for feature.
+- added possibility to create bloc and cubit template.
+- added possibility to add packages to a project.
+
 ## 0.1.0
 - create and Connect commands have been changed.
 - added possibility to generate keystore file and connect it to the project.
 
 ## 0.0.1
 - added possibility to create a Flutter project template.
-- added the ability to connect a remote Github repository.
+- added possibility to connect a remote Github repository.